Set dtype to uint8 on filled mask data.

+++ datacube_ows/data.py
@@ -267,7 +267,7 @@ class DataStacker:
                 for var in data.data_vars.variables.keys():
                     break
                 template = getattr(data, var)
-                new_data = numpy.ndarray(template.shape)
+                new_data = numpy.ndarray(template.shape, dtype="uint8")
                 new_data.fill(0)
                 qry_result = template.copy(data=new_data)
                 data_new_bands = {}
